requirements.txt
pytest==8.3.5
allure-pytest==2.15.0